requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==3.0.4
click==7.1.2
dnacentersdk==2.0.2
fastjsonschema==2.15.1
Flask==2.0.1
future==0.18.2
idna==2.10
itsdangerous>=2.0
Jinja2==3.0.1
MarkupSafe>=2.0
meraki==1.4.3
multidict==5.1.0
PyJWT==1.7.1
python-dotenv==0.14.0
requests==2.25.1
requests-toolbelt==0.9.1
typing-extensions==3.10.0.0
urllib3==1.26.5
webexteamssdk==1.6
Werkzeug==2.0.1
yarl==1.6.3
